==> Data analysis/12 - Lags.Rmd <==
---
title: "Lags"
author: "Abby Lewis"
date: "2023-05-22"
output: html_document
---

This file analyzes relevant lags for ABA relationships of interest. 

Table of contents:
Step 1: Load packages and data
Step 2: Create function to calculate lags and plot
Step 3: Generate all figures


Step 1: Load packages and data
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)

set.seed(47) #Set seed so points jitter the same every time

#Load saved data
with_temp <- read.csv("../Compiled data/All_data_annual AF.csv")
```


Step 2: Create function to calculate lags and plot
```{r}
plot_lags <- function(response_var, 
                      driver_var, 
                      data, ylab, 
                      xlab) {
  
  #Parse variable names
  response <- sym(response_var)
  driver <- sym(driver_var)
  
  #Calculate correlations
  driver_response_stat <- data %>%
    rename(driver = !!driver) %>%
    unique() %>%
    group_by(LakeID) %>%
    arrange(LakeID, Year) %>%
    mutate(driver_lag := lag(driver),
           driver_lag2 = lag(driver_lag),
           driver_lag2 := ifelse(lag(lag(Year)) == (Year - 2),
                                 driver_lag2,
                                 NA),
           lag_is_last_year = ifelse(lag(Year) == (Year - 1), T, F)) %>%
    filter(lag_is_last_year) %>%
    pivot_longer(cols = c(driver, 
                          driver_lag, 
                          driver_lag2)) %>%
    filter(!is.na(value), !is.na(!!response)) %>%
    group_by(name, LakeID) %>%
    mutate(nyear = length(unique(Year))) %>%
    arrange(LakeID) %>%
    filter(nyear >= 10) %>%
    dplyr::summarize(driver_response = cor.test(!!response, 
                                                value, 
                                                method = "spearman")$estimate)
  
  #Plot
  plot = driver_response_stat%>%
    group_by(name)%>%
    mutate(p = wilcox.test(driver_response)$p.value,
           n = n())%>%
    ggplot(aes(x=factor(name, 
                        levels = c("driver_lag2","driver_lag","driver"), 
                        labels = c("t-2","t-1","t")), 
               y = driver_response))+
    geom_hline(yintercept = 0)+
    geom_boxplot(aes(color = p < 0.05), outlier.shape = NA)+
    geom_jitter(width = 0.15, color = "grey20", fill = "white", shape = 21)+
    geom_text(aes(label = paste0("n = ",n), 
                  y = max(driver_response, na.rm = T) + 0.1))+
    ylab(ylab)+
    scale_color_manual(values = c("grey50","#FB8B24"), 
                       breaks = c(F,T),
                       labels = c("FALSE","TRUE"),
                       drop = F, 
                       name = "p < 0.05")+
    xlab(xlab)+
    theme_bw()
  
  return(plot)
}
```


Step 3: Generate all figures
```{r}
jpeg("../Figures/Hypo P vs DO lag.jpg", width = 4, height = 3.5, res = 300, units = "in")
plot_lags(driver = "DO_mgL_HYPO", 
          response = "TP_ugL_HYPO", 
          data = with_temp,
          ylab = "Correlation between hypo. DO and hypo. TP",
          xlab = "Hypo. DO lag")
dev.off()

jpeg("../Figures/Hypo P vs AF lag.jpg", width = 4, height = 3.5, res = 300, units = "in")
plot_lags(driver = "AF", 
          response = "TP_ugL_HYPO", 
          data = with_temp,
          ylab = "Correlation between AF and hypo. TP",
          xlab = "Anoxic factor lag")
dev.off()

jpeg("../Figures/Epi P vs hypo P lag.jpg", width = 4, height = 3.5, res = 300, units = "in")
plot_lags(driver = "TP_ugL_HYPO", 
          response = "TP_ugL_EPI", 
          data = with_temp,
          ylab = "Correlation between hypo. TP and epi. TP",
          xlab = "Hypo. TP lag")
dev.off()

jpeg("../Figures/Epi P vs chl-a lag.jpg", width = 4, height = 3.5, res = 300, units = "in")
plot_lags(driver = "strat_TP_ugL_EPI", 
          response = "Chla_ugL_EPI", 
          data = with_temp,
          ylab = "Correlation between epi. TP and chl-a",
          xlab = "Epi. TP lag")
dev.off()

jpeg("../Figures/Chla vs Demand lag.jpg", width = 4, height = 3.5, res = 300, units = "in")
plot_lags(driver = "Chla_ugL_EPI", 
          response = "DO_demand_mgLd_HYPO", 
          data = with_temp,
          ylab = "Correlation between epi. chl-a and VHOD",
          xlab = "Epi. chl-a lag")
dev.off()

jpeg("../Figures/Hypo. DO vs vhod lag.jpg", width = 4, height = 3.5, res = 300, units = "in")
plot_lags(driver = "DO_demand_mgLd_HYPO", 
          response = "DO_mgL_HYPO", 
          data = with_temp,
          ylab = "Correlation between VHOD and hypo. DO",
          xlab = "VHOD lag")
dev.off()

jpeg("../Figures/AF vs vhod lag.jpg", width = 4, height = 3.5, res = 300, units = "in")
plot_lags(driver = "DO_demand_mgLd_HYPO", 
          response = "AF", 
          data = with_temp,
          ylab = "Correlation between VHOD and AF",
          xlab = "VHOD lag")
dev.off()
```
