==> Data analysis/13 - Regressions.Rmd <==
---
title: "Mixed effect regressions"
author: "Abby Lewis"
date: "2023-05-22"
output: html_document
---

This file runs the model selection algorithm and regressions both within and across lakes

Table of contents:
Step 1: Load packages and data
Step 2: Format data for regressions
Step 3: Model selection
Step 4: Segmented TP regression
Step 5: Generate combined figures
Step 6: Driver analysis
Step 7: Before and after the first year of anoxia (figure)


Step 1: Load packages and data
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(MuMIn)
library(car)
library(lme4)
library(ggpubr)
library(ggridges)
library(gtable)
library(gridExtra)
library(grid)
library(ggh4x)
library(segmented)
library(ggpmisc)
library(tidyverse)
library(lubridate)
source("lmer_functions.R")

POSTER = F #Set global option: Should figures be designed for a poster or for the manuscript?

set.seed(47) #Set seed so points jitter the same way each time

#Load saved data
with_temp <- read.csv("../Compiled data/All_data_annual AF.csv")
```


Step 2: Format data for regressions
```{r}
cummean.na <- function(x, na.rm = T) {
  n <- length(x)
  op <- rep(NA, n)
  for(i in 1:n) {op[i] <- mean(x[1:i], na.rm = !!na.rm)}
  rm(x, na.rm, n, i)
  return(op)
}

#Add relevant lags and seasonal calculations
with_lags <- with_temp%>%
  unique()%>%
  group_by(LakeID)%>%
  arrange(LakeID,Year)%>%
  mutate(chla_lag = lag(Chla_ugL_EPI),
         epi_p_lag = lag(TP_ugL_EPI),
         strat_epi_p_lag = lag(strat_TP_ugL_EPI),
         hypo_p_lag = lag(TP_ugL_HYPO),
         epi_n_lag = lag(TN_ugL_EPI),
         hypo_n_lag = lag(TN_ugL_HYPO),
         AF_lag = lag(AF),
         summer_temp = (temp_july + temp_aug)/2,
         spring_temp = (temp_mar + temp_apr)/2,
         winter_temp = (temp_jan + temp_feb)/2,
         fall_temp = (temp_sep+temp_oct)/2,
         summer_precip = (precip_july+ precip_aug)/2,
         spring_precip = (precip_mar + precip_apr)/2,
         winter_precip = (precip_jan + precip_feb)/2,
         anoxic = max(DO_mgL_HYPO,na.rm = T) < 1.8,
         lag_is_last_year = ifelse(lag(Year) == (Year-1), T, F))%>%
  mutate(across(contains("_lag"), function(x) ifelse(lag_is_last_year, x, NA)))

# Save this as the dataset we will work with
dataset <- with_lags
row.names(dataset) <- paste(dataset$LakeID, dataset$Year)

data_log_nuts <- dataset%>%
  mutate(Chla_ugL_EPI=log(Chla_ugL_EPI),
         chla_lag = log(chla_lag),
         TP_ugL_EPI = log(TP_ugL_EPI),
         strat_TP_ugL_EPI = log(strat_TP_ugL_EPI),
         TP_ugL_HYPO = log(TP_ugL_HYPO),
         TN_ugL_EPI = log(TN_ugL_EPI),
         strat_TN_ugL_EPI = log(strat_TN_ugL_EPI),
         TN_ugL_HYPO = log(TN_ugL_HYPO),
         epi_p_lag = log(epi_p_lag),
         strat_epi_p_lag = log(strat_epi_p_lag),
         hypo_p_lag = log(hypo_p_lag),
         epi_n_lag = log(epi_n_lag),
         hypo_n_lag = log(hypo_n_lag)) #produces NAs because of two EPI TN == 0

#Create datasets that are only oxic or only anoxic lakes
data_always_oxic <- dataset %>%
  group_by(LakeID) %>%
  filter(!is.finite(max(AF, na.rm = T)) || max(AF, na.rm = T) > 0,
         min(DO_mgL_HYPO, na.rm = T) > 1.8)

data_no_oxic <- dataset %>%
  group_by(LakeID) %>%
  filter(is.finite(max(AF, na.rm = T)) && max(AF, na.rm = T) > 0)

#How many always oxic lakes are we removing?
compare <- dataset %>%
  group_by(LakeID) %>%
  filter(is.finite(max(AF, na.rm = T)))
length(unique(compare$LakeID)) - length(unique(data_no_oxic$LakeID))
```


Step 3: Model selection
```{r}
responses <- c("DO_demand_mgLd_HYPO")
potential_drivers <- c("Chla_ugL_EPI",
                      "chla_lag",
                      "strat_Temp_C_HYPO",
                      #"DOC_mgL_HYPO", #Adding this limits data availability, making effects more difficult to detect. HOWEVER this does not come out as important
                      "SA_vol_ratio_HYPO",
                      "strat_buoyancy_freq_EPI" #all layers are the same
                      )
aic_calculator_lmer(data_log_nuts,responses,potential_drivers, interaction = "+")
selected_drivers <- c("Chla_ugL_EPI","chla_lag","strat_Temp_C_HYPO","SA_vol_ratio_HYPO")
std_data <- standardize_data(data_log_nuts,responses,selected_drivers)
all_lakes_vhod <- mod_by_lake(data_log_nuts,responses,selected_drivers, interaction = "+")
mod_lmer <- lmer(DO_demand_mgLd_HYPO~Chla_ugL_EPI+chla_lag+strat_Temp_C_HYPO+SA_vol_ratio_HYPO+(1|LakeID), data = std_data)
mod_lmer
vif(mod_lmer)
vhod <- plot_effects_lmer(mod_lmer,"Oxygen demand (mg/L/d)", poster = POSTER)
vhod_ridge <- plot_effects_by_lake_lmer_ridge(all_lakes_vhod,var_name = "Oxygen demand (mg/L/d)",mod_lmer, poster = POSTER)

### DO (only oxic lakes)
responses <- c("DO_mgL_HYPO")
potential_drivers <- c("DO_demand_mgLd_HYPO",
                      "spring_temp",
                      "summer_temp",
                      "winter_temp",
                      "Temp_C_HYPO",
                      "DO_date_HYPO"
                      )
aic_calculator_lmer(data_always_oxic,responses,potential_drivers, interaction = "+")
selected_drivers <- c("DO_demand_mgLd_HYPO", "spring_temp", "DO_date_HYPO")
std_data <- standardize_data(data_always_oxic,responses,selected_drivers)
all_lakes_do <- mod_by_lake(data_always_oxic,responses,selected_drivers, interaction = "+")
mod_lmer <- lmer(DO_mgL_HYPO~DO_demand_mgLd_HYPO+spring_temp+DO_date_HYPO+(1|LakeID), data = std_data) 
vif(mod_lmer)
hypo_do_oxic <- plot_effects_lmer(mod_lmer,"Hypolimnetic DO (in oxic lakes)", poster = POSTER) 
hypo_do_oxic_ridge <- plot_effects_by_lake_lmer_ridge(all_lakes_do,"Hypolimnetic DO (in oxic lakes)",mod_lmer, poster = POSTER)

### Anoxic factor
responses <- c("AF")
potential_drivers <- c("DO_demand_mgLd_HYPO",
                      "spring_temp",
                      "fall_temp",
                      "winter_temp",
                      "Temp_C_HYPO"
                      )
aic_calculator_lmer(dataset,responses,potential_drivers, interaction = "+")
aic_calculator_lmer(data_no_oxic,responses,potential_drivers, interaction = "+") #same results - using no oxic
selected_drivers <- c("DO_demand_mgLd_HYPO","spring_temp","fall_temp","Temp_C_HYPO")
std_data <- standardize_data(dataset,responses,selected_drivers) #all lakes for across-lake regression
all_lakes_af <- mod_by_lake(data_no_oxic,responses,selected_drivers, interaction = "+") #only oxic lakes here
mod_lmer <- lmer(AF~spring_temp+DO_demand_mgLd_HYPO+Temp_C_HYPO+fall_temp+(1|LakeID), data = std_data) 
mod_lmer
vif(mod_lmer)
af <- plot_effects_lmer(mod_lmer,"Anoxic factor", poster = POSTER) 
af_ridge <- plot_effects_by_lake_lmer_ridge(all_lakes_af,"Anoxic factor",mod_lmer, poster = POSTER)

### VHOD5
responses <- c("VHOD5_mgLd_HYPO")
potential_drivers <- c("Chla_ugL_EPI",
                      "chla_lag",
                      "spring_temp",
                      "strat_Temp_C_HYPO",
                      "strat_buoyancy_freq_EPI"
                      )
aic_calculator_lmer(data_log_nuts,responses,potential_drivers, interaction = "+")
selected_drivers <- c("Chla_ugL_EPI","chla_lag","strat_buoyancy_freq_EPI")
std_data <-standardize_data(data_log_nuts,responses,selected_drivers)
all_lakes_vhod5 <- mod_by_lake(data_log_nuts,responses,selected_drivers, interaction = "+")
mod_lmer <- lmer(VHOD5_mgLd_HYPO~Chla_ugL_EPI+chla_lag+strat_buoyancy_freq_EPI+(1|LakeID), data = std_data) 
vif(mod_lmer)
confint(mod_lmer)
summary(mod_lmer)
plot_effects_lmer(mod_lmer,"Temperature-corrected oxygen demand") 

jpeg("../Figures/MLR-lmer/Parameter estimate-Temperature-corrected oxygen demand-all_lakes-ridge wide.jpg", width = 4.5, height = 4, res = 300, units = "in")
plot_effects_by_lake_lmer_ridge(all_lakes_vhod5,"Temperature-corrected oxygen demand",mod_lmer)
dev.off()

### Epi. TP
responses <- c("strat_TP_ugL_EPI")
potential_drivers <- c("TP_ugL_HYPO",
                      "hypo_p_lag",
                      "epi_p_lag",
                      "spring_precip",
                      "summer_precip",
                      "winter_precip",
                      "strat_buoyancy_freq_EPI",
                      "strat_TP_date_EPI"
                      )
aic_calculator_lmer(data_log_nuts,responses,potential_drivers, interaction = "+")
selected_drivers <- c("TP_ugL_HYPO","hypo_p_lag","epi_p_lag","spring_precip","summer_precip","strat_TP_date_EPI")
std_data <- standardize_data(data_log_nuts,responses,selected_drivers)
all_lakes_epiP <- mod_by_lake(data_log_nuts,responses,selected_drivers, interaction = "+")
mod_lmer <- lmer(strat_TP_ugL_EPI~TP_ugL_HYPO+hypo_p_lag+epi_p_lag+spring_precip+summer_precip+strat_TP_date_EPI+(1|LakeID), data = std_data)
mod_lmer
vif(mod_lmer)
epi_p <- plot_effects_lmer(mod_lmer,"Epilimnetic TP (µg/L)", poster = POSTER) 
epi_p_ridge <- plot_effects_by_lake_lmer_ridge(all_lakes_epiP,"Epilimnetic TP (µg/L)",mod_lmer, poster = POSTER)

### TP with AF
responses <- c("TP_ugL_HYPO")
potential_drivers <- c("AF",
                       "strat_TP_ugL_EPI",
                       "strat_buoyancy_freq_EPI",
                       "strat_Temp_C_HYPO",
                       "spring_precip",
                       "summer_precip",
                       "winter_precip",
                       "TP_date_HYPO"
                       )
aic_calculator_lmer(data_log_nuts,responses,potential_drivers, interaction = "+")
selected_drivers <- c("strat_TP_ugL_EPI","AF", "winter_precip")
std_data <-standardize_data(data_log_nuts,responses,selected_drivers)
all_lakes_hypoP_af <- mod_by_lake(data_log_nuts,responses,selected_drivers, interaction = "+")
mod_lmer <- lmer(TP_ugL_HYPO~strat_TP_ugL_EPI+AF+winter_precip+(1|LakeID), data = std_data)
vif(mod_lmer)
mod_lmer
hypo_p_af <- plot_effects_lmer(mod_lmer,"Hypolimnetic TP (µg/L)", poster = POSTER) 
hypo_p_ridge_af <- plot_effects_by_lake_lmer_ridge(all_lakes_hypoP_af,"Hypolimnetic TP (µg/L)",mod_lmer, poster = POSTER)

### Epi. chl-a
responses <- c("Chla_ugL_EPI")
potential_drivers <- c("strat_TP_ugL_EPI",
                      "strat_TN_ugL_EPI", 
                      "spring_temp",
                      "summer_temp",
                      "Chla_date_EPI")

length(unique(data_log_nuts$LakeID[!is.na(data_log_nuts$strat_TN_ugL_EPI)])) #86
length(unique(data_log_nuts$LakeID[!is.na(data_log_nuts$strat_TP_ugL_EPI)])) #387

aic_calculator_lmer(data_log_nuts,responses,potential_drivers, interaction = "+")
selected_drivers <- c("spring_temp","strat_TP_ugL_EPI")
std_data <- standardize_data(data_log_nuts,responses,selected_drivers)
all_lakes_chla <- mod_by_lake(data_log_nuts,responses,selected_drivers, interaction = "+")
mod_lmer <- lmer(Chla_ugL_EPI~spring_temp+strat_TP_ugL_EPI+(1|LakeID), data = std_data) 
mod_lmer
vif(mod_lmer)
epi_chla <- plot_effects_lmer(mod_lmer,"Epilimnetic chl-a (µg/L)", poster = POSTER) 
epi_chla_ridge <- plot_effects_by_lake_lmer_ridge(all_lakes_chla,"Epilimnetic chl-a (µg/L)",mod_lmer, poster = POSTER)
```


Step 4: Segmented TP regression
```{r}
### Without AF
responses <- c("TP_ugL_HYPO")
potential_drivers <- c("DO_mgL_HYPO",
                      "strat_TP_ugL_EPI",
                      "strat_buoyancy_freq_EPI",
                      "strat_Temp_C_HYPO",
                      "spring_precip",
                      "summer_precip",
                      "winter_precip",
                      "TP_date_HYPO"
                      )
aic_calculator_lmer(data_log_nuts,responses,potential_drivers, interaction = "+")
selected_drivers <- c("DO_mgL_HYPO","strat_TP_ugL_EPI","winter_precip")
#SEGMENTED
mod_seg <- lme(TP_ugL_HYPO~DO_mgL_HYPO+strat_TP_ugL_EPI+winter_precip, random=~1|LakeID, data = data_log_nuts%>%
  dplyr::select(all_of(c(selected_drivers,responses,"LakeID")))%>%
  filter(if_all(where(is.numeric),is.finite))) #for some reason lmer doesn't work here
seglme <- segmented.lme(mod_seg, random=list(LakeID=pdDiag(~1)), seg.Z = ~DO_mgL_HYPO, psi = 1)
#G0 is the breakpoint, U is added to the coefficient for DO_mgL_HYPO to get the slope after the breakpoint
break_untrans <- round(unique(seglme$fixed.eta.psi),1)
summary(seglme)

png("../Figures/DO TP black.png", width = 3, height = 3, res = 1000, units = "in")
do_tp <- with_lags%>%
  filter(!is.na(TP_ugL_HYPO))%>%
  ungroup()%>%
  mutate(label = paste0("n = ", n(), " lake-years\nn = ",
                        length(unique(LakeID))," lakes"),
         on_other_fig = ifelse(!is.na(TP_ugL_HYPO) & 
                                 !is.na(depth_ratio_HYPO) & 
                                 !is.na(DO_mgL_HYPO) & 
                                 DO_mgL_HYPO < 1.8 & 
                                 depth_ratio_HYPO < 1,
                               "y",
                               "n"))%>%
  ggplot()+
  scale_y_log10()+
  geom_point(aes(x = DO_mgL_HYPO,y=TP_ugL_HYPO,color = on_other_fig), alpha = 0.2)+
  geom_text(aes(x = 2.5, y=5000,label = label),data = .%>%dplyr::select(label)%>%unique(), hjust = 0, size = 3.5)+
  xlab("Hypolimnetic DO (mg/L)")+
  ylab("Hypolimnetic TP (µg/L)")+
  scale_color_manual(values = c("black","black"))+
  geom_vline(xintercept = 1.8, color = "black")+
  theme_bw()+
  theme(legend.position = "none")
do_tp
dev.off()
```


Step 5: Generate combined figures
```{r}
g1 <- ggplotGrob(
  hypo_p_af  +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9),
                         axis.title.x = element_blank())+
                   force_panelsizes(rows = unit(3*5/6, "cm")))
g2 <- ggplotGrob(
  epi_p   +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9),
                         axis.title.x = element_blank())+
                   force_panelsizes(rows = unit(6*5/6, "cm")))
g3 <- ggplotGrob(
  epi_chla+theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9))+
                   force_panelsizes(rows = unit(2*5/6, "cm")))
g4 <- ggplotGrob(
  vhod    +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9),
                         axis.title.x = element_blank())+
                   force_panelsizes(rows = unit(4*5/6, "cm")))
g5 <- ggplotGrob(
  af +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9))+
                   force_panelsizes(rows = unit(4*5/6, "cm")))
maxWidth <- unit.pmax(g1$widths,g2$widths,g3$widths,g4$widths,g5$widths)
g1$widths <- maxWidth
g2$widths <- maxWidth
g3$widths <- maxWidth
g4$widths <- maxWidth
g5$widths <- maxWidth

layout <- rbind(c(1, 4),
                c(1, 4),
                c(1, 4),
                c(1, 4),
                c(1, 4),
                c(2, 4),
                c(2, 5),
                c(2, 5),
                c(2, 5),
                c(2, 5),
                c(2, 5),
                c(2, 5),
                c(2, 5),
                c(3, NA),
                c(3, NA),
                c(3, NA),
                c(3, NA),
                c(3, NA))
jpeg("../Figures/MLR-lmer/Parameter estimate-all-custom layout-AF.jpeg", width = 7, height = 6, res = 600, units = "in")
grid.arrange(g1,g2,g3,g4,g5,layout_matrix=layout)
dev.off()

g1 <- ggplotGrob(hypo_p_ridge_af  
                 #+ggtitle("Bottom-water phosphorus")
                 +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9),
                         axis.title.x = element_blank())+
                   force_panelsizes(rows = unit(3, "cm")))
g2 <- ggplotGrob(epi_p_ridge  
                 #+ggtitle("Surface phosphorus")
                 +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9),
                         axis.title.x = element_blank())+
                   force_panelsizes(rows = unit(6, "cm")))
g3 <- ggplotGrob(epi_chla_ridge
                 #+ggtitle("Surface cholorphyll-a")
                 +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9),
                         axis.title.x = element_blank())+
                   force_panelsizes(rows = unit(2, "cm")))
g4 <- ggplotGrob(vhod_ridge    
                 #+ggtitle("Oxygen demand")
                 +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9))+
                   force_panelsizes(rows = unit(4, "cm")))
g5 <- ggplotGrob(af_ridge 
                 #+ggtitle("Bottom-water oxygen")
                 +theme(plot.title = element_text(size = 10),
                         plot.subtitle = element_text(size = 9))+
                   force_panelsizes(rows = unit(4, "cm")))
maxWidth <- unit.pmax(g1$widths,g2$widths,g3$widths,g4$widths,g5$widths)
g1$widths <- maxWidth
g2$widths <- maxWidth
g3$widths <- maxWidth
g4$widths <- maxWidth
g5$widths <- maxWidth
layout <- rbind(c(1, 2),
                c(1, 2),
                c(1, 2),
                c(1, 2),
                c(1, 2),
                c(1, 2),
                c(1, 2),
                c(1, 2),
                c(3, 2),
                c(3, 2),
                c(3, 2),
                c(3, 2),
                c(3, 2),
                c(3, 5),
                c(3, 5),
                c(4, 5),
                c(4, 5),
                c(4, 5),
                c(4, 5),
                c(4, 5),
                c(4, 5),
                c(4, 5),
                c(4, 5),
                c(4, 5),
                c(4, NA),
                c(4, NA))
jpeg("../Figures/MLR-lmer/Parameter estimate-all-ridge-af.jpeg", width = 6, height = 6, res = 600, units = "in")
grid.arrange(g1,g2,g3,g4,g5, layout_matrix=layout)
dev.off()
```


Step 6: Driver analysis
```{r}
incomplete <- all_lakes_chla%>%
  dplyr::select(strat_TP_ugL_EPI,LakeID)%>%
  full_join(all_lakes_epiP%>%
              dplyr::select(TP_ugL_HYPO, hypo_p_lag,LakeID))%>%
  full_join(all_lakes_do%>%
              dplyr::select(DO_demand_mgLd_HYPO,spring_temp,LakeID))%>%
  full_join(all_lakes_vhod%>%
              dplyr::select(chla_lag,Chla_ugL_EPI,LakeID))%>%
  full_join(all_lakes_hypoP_af%>%
              dplyr::select(AF, LakeID))%>%
  full_join(all_lakes_af%>%
              dplyr::select(DO_demand_mgLd_HYPO, LakeID)%>%
              rename(DO_demand_mgLd_HYPO_af = DO_demand_mgLd_HYPO))

hydro <- read.csv("../Compiled data/hydrolakes_full.csv")
lat_long <- read.csv("https://pasta.lternet.edu/package/data/eml/edi/1530/1/fadd3eaa25b5fdd1fc4efba70e660579")

means <- with_temp%>%
  group_by(LakeID)%>%
  dplyr::summarize(mean_do = mean(DO_mgL_HYPO,na.rm=T),
                   median_do = median(DO_mgL_HYPO,na.rm=T),
                   mean_log_do = mean(log(DO_mgL_HYPO),na.rm=T),
                   sd_do = sd(DO_mgL_HYPO,na.rm=T),
                   sd_log_do = sd(log(DO_mgL_HYPO),na.rm=T),
                   cv_log_do = unique(sd_log_do)/unique(mean_log_do),
                   max_do = max(DO_mgL_HYPO, na.rm=T),
                   min_do = min(DO_mgL_HYPO, na.rm = T),
                   cv_do = unique(sd_do)/unique(mean_do),
                   mean_hypo_p = mean(TP_ugL_HYPO,na.rm=T),
                   log_sd_hypo_p = log(sd(TP_ugL_HYPO,na.rm=T)),
                   cv_hypo_p = sd(TP_ugL_HYPO,na.rm=T)/unique(mean_hypo_p),
                   mean_epi_p = mean(TP_ugL_EPI,na.rm=T),
                   sd_epi_p = sd(TP_ugL_EPI,na.rm=T),
                   cv_epi_p = unique(sd_epi_p)/unique(mean_epi_p),
                   mean_epi_n = mean(TN_ugL_EPI,na.rm=T),
                   mean_demand = mean(DO_demand_mgLd_HYPO,na.rm=T),
                   sd_demand = sd(DO_demand_mgLd_HYPO,na.rm=T),
                   cv_demand = unique(sd_demand)/unique(mean_demand),
                   mean_chla = mean(Chla_ugL_EPI,na.rm=T),
                   sd_chla = sd(Chla_ugL_EPI,na.rm=T),
                   cv_chla = unique(sd_chla)/unique(mean_chla),
                   mean_sa_vol = mean(SA_vol_ratio_HYPO),
                   mean_hypo_temp = mean(Temp_C_HYPO,na.rm=T),
                   mean_epi_temp = mean(Temp_C_EPI,na.rm=T),
                   mean_air_temp = mean(mean_temp, na.rm=T),
                   mean_bf = mean(strat_buoyancy_freq_EPI,na.rm=T),
                   mean_af = mean(AF, na.rm = T))

for_mlr <- incomplete%>%
  left_join(lat_long, by = c("LakeID"))%>%
  left_join(means)%>%
  left_join(hydro%>%
              dplyr::select(ID,Shore_dev,Res_time,Wshd_area,Lake_type), by = c("LakeID"="ID"))%>%
  mutate(log_MaximumDepth_m = log(MaximumDepth_m),
         log_SurfaceArea_ha = log(SurfaceArea_ha),
         log_MeanDepth_m = log(MeanDepth_m),
         log_Volume = log(MeanDepth_m*SurfaceArea_ha),
         log_Res_time= log(Res_time),
         log_epiP = log(mean_epi_p),
         log_hypoP = log(mean_hypo_p),
         log_af = log(mean_af),
         SA_depth = log_SurfaceArea_ha/log_MaximumDepth_m,
         Lake_type = factor(Lake_type),
         tn_tp_molar_ratio = (mean_epi_n/14.01)/(mean_epi_p/30.97),
         log_dynamic_ratio = log(MeanDepth_m/sqrt(SurfaceArea_ha)))

potential_drivers <- c("log_MeanDepth_m",
                      "log_MaximumDepth_m",
                      "log_SurfaceArea_ha",
                      "log_dynamic_ratio",
                      "log_Res_time",
                      "Elevation_m",
                      "log_hypoP"
                      )
responses <- c("TP_ugL_HYPO") #Nothing comes out cleanly
#responses <- c("hypo_p_lag") #Nothing comes out cleanly if we remove one outlier %>%filter(hypo_p_lag>-1)
aic_calculator_onevar(for_mlr,responses,potential_drivers)
summary(lm(TP_ugL_HYPO~log_Res_time, data = for_mlr)) 

#Hypo P lag shows expected pattern with res time but driven by 1 point
for_mlr%>%
  filter(!is.na(hypo_p_lag))%>%
  ggplot(aes(x=Res_time,y=hypo_p_lag))+
  geom_point()+
  scale_x_log10()+
  geom_hline(yintercept = 0)+
  geom_smooth(method = "lm", color = "black")+
  theme_bw()+
  ylab("Coefficient for the effect\nof hypo. TP (t-1) on epi TP")+
  xlab("Residence time (d)")

potential_drivers <- c("log_MeanDepth_m",
                      "log_MaximumDepth_m",
                      "log_SurfaceArea_ha",
                      "log_dynamic_ratio",
                      "log_Res_time",
                      "Elevation_m",
                      "log_epiP"
                      #"tn_tp_ratio" #removing for the sake of having a standardized approach
                      )
responses <- c("strat_TP_ugL_EPI")
aic_calculator_onevar(for_mlr,responses,potential_drivers)
summary(lm(strat_TP_ugL_EPI~log_epiP, data = for_mlr)) #Take home: epi P has a positive relationship

p_value = summary(lm(strat_TP_ugL_EPI~log_epiP, data = for_mlr))$coefficients[,4][2]
if(p_value < 0.001){p_value_text = "< 0.001"}else{p_value_text = paste0("= ",as.character(p_value))}
r2_value = summary(lm(strat_TP_ugL_EPI~log_epiP, data = for_mlr))$r.squared
r2_value_text = round(r2_value,2)

jpeg("../Figures/Effect of TP on chl-a.jpg", width = 4, height = 4, res = 300, units = "in")
chla <- for_mlr%>%
  filter(!is.na(strat_TP_ugL_EPI))%>%
  ggplot(aes(x=mean_epi_p,y=strat_TP_ugL_EPI))+
  geom_point(color = "grey50")+
  geom_smooth(method = "lm", color = "black")+
  scale_x_log10()+
  geom_hline(yintercept = 0)+
  theme_bw()+
  ylab("Coefficient for the effect\nof epi. TP on chl-a")+
  xlab("Mean epilimnetic TP (µg/L)")+
  annotate("text", x = 8, y = 0.87,
           label = paste0("R^2 == ",r2_value_text,"~~~p ", p_value_text), 
           parse = TRUE)
chla
dev.off()

jpeg("../Figures/Effect of TP on chl-a - TNTP ratio.jpg", width = 6, height = 4, res = 300, units = "in")
for_mlr%>%
  filter(!is.na(strat_TP_ugL_EPI))%>%
  mutate(ratio = ifelse(tn_tp_molar_ratio>110,"P limited (TN:TP > 110)",
                        ifelse(tn_tp_molar_ratio<44,"N limited (TN:TP < 44)","Co-limited")),
         ratio = ifelse(is.na(ratio),"Insufficient data",ratio),
         ratio = factor(ratio,levels = c("N limited (TN:TP < 44)","Co-limited","P limited (TN:TP > 110)","Insufficient data")))%>%
  ggplot(aes(x = mean_epi_p, y = strat_TP_ugL_EPI, color = ratio))+
  geom_smooth(method = "lm", color = "black")+
  geom_point()+
  scale_x_log10()+
  geom_hline(yintercept = 0)+
  theme_bw()+
  theme(legend.title = element_blank())+
  ylab("Coefficient for the effect\nof epi. TP on chl-a")+
  xlab("Mean epilimnetic TP (µg/L)")
dev.off()

jpeg("../Figures/TP mean and SD.jpg", width = 4, height = 4, res = 300, units = "in")
for_mlr%>%
  filter(!is.na(mean_epi_p),
         !is.na(sd_epi_p))%>%
  ggplot(aes(x=mean_epi_p,y=sd_epi_p))+
  geom_point()+
  scale_x_log10()+
  scale_y_log10()+
  theme_bw()+
  ylab("SD")+
  xlab("Mean epilimnetic TP (µg/L)")+
  ylab("SD epilimnetic TP (µg/L)")
dev.off()

potential_drivers <- c("log_MeanDepth_m",
                      "log_MaximumDepth_m",
                      "log_SurfaceArea_ha",
                      "log_dynamic_ratio",
                      "log_Res_time",
                      "Elevation_m",
                      "log_af"
                      )
responses <- c("AF")
aic_calculator_onevar(for_mlr,responses,potential_drivers)
summary(lm(AF~log_Res_time, data = for_mlr)) #Take home: nothing comes out cleanly

potential_drivers <- c("log_MeanDepth_m",
                      "log_MaximumDepth_m",
                      "log_SurfaceArea_ha",
                      "log_Res_time",
                      "log_dynamic_ratio",
                      "Elevation_m",
                      "mean_demand"
                      )
responses <- c("DO_demand_mgLd_HYPO_af")
aic_calculator_onevar(for_mlr,responses,potential_drivers)
summary(lm(DO_demand_mgLd_HYPO_af~log_MeanDepth_m, data = for_mlr)) #Take home: mean depth is most important

p_value = summary(lm(DO_demand_mgLd_HYPO_af~log_MeanDepth_m, 
                     data = for_mlr #%>% filter(MeanDepth_m > 1)
                     ))$coefficients[,4][2]
if(p_value < 0.001){p_value_text = "< 0.001"}else{
  p_value_text = paste0("== ",as.character(round(p_value,3)))
  }
r2_value = summary(lm(DO_demand_mgLd_HYPO_af~log_MeanDepth_m, 
                      data = for_mlr #%>% filter(MeanDepth_m > 1)
                      ))$r.squared
r2_value_text = round(r2_value,2)

# Identify influential points
row.names(for_mlr) <- for_mlr$LakeID
model <- lm(DO_demand_mgLd_HYPO_af ~ log10(MeanDepth_m), data = for_mlr)
cooksD <- cooks.distance(model)
influential <- cooksD[(cooksD > (3 * mean(cooksD, na.rm = TRUE)))]
influential_lakes <- names(influential)

summary(lm(DO_demand_mgLd_HYPO_af ~ log10(MeanDepth_m), data = for_mlr))
summary(lm(DO_demand_mgLd_HYPO_af ~ log10(MeanDepth_m), data = for_mlr %>% filter(!LakeID %in% influential_lakes)))
length(unique(influential_lakes))

jpeg("../Figures/Effect of VHOD on AF.jpg", width = 4, height = 4, res = 300, units = "in")
demand_af <- for_mlr%>%
  filter(!is.na(DO_demand_mgLd_HYPO_af),
         !is.na(MeanDepth_m))%>%
  ggplot(aes(x = MeanDepth_m, y = DO_demand_mgLd_HYPO_af))+
  geom_point(color = "grey50")+
  geom_smooth(method = "lm", color = "black")+
  geom_smooth(method = "lm", color = "black", lty = "31",
              data = . %>% filter(!LakeID %in% influential_lakes))+
  scale_x_log10()+
  geom_hline(yintercept = 0)+
  theme_bw()+
  #scale_color_viridis_c(trans = "log", breaks = c(25,50,100,200), name = "Maximum\ndepth (m)")+
  ylab("Coefficient for the effect\nof VHOD on AF")+
  xlab("Mean depth (m)")+
  guides(shape = "none") +
  annotate("text",x = 3, y = 1.15,
           label = paste0("R^2 == ",r2_value_text,"~~~p ", p_value_text), 
           parse = TRUE)
demand_af
dev.off()

jpeg("../Figures/VHOD mean and depth.jpg", width = 4, height = 4, res = 300, units = "in")
for_mlr%>%
  filter(!is.na(mean_demand),
         !is.na(MeanDepth_m))%>%
  ggplot(aes(x=MeanDepth_m,y=mean_demand))+
  geom_point()+
  geom_smooth(method = "lm", color = "black")+
  scale_x_log10()+
  #scale_y_log10()+
  theme_bw()+
  ylab("SD")+
  xlab("Mean depth (m)")+
  ylab("Mean oxygen demand (mg/L/d)")+
  annotate("text",x = 3, y = 0.5,
           label = paste0("R^2 == ",r2_value_text,"~~~p ", p_value_text), 
           parse = TRUE)
dev.off()

potential_drivers <- c("log_MeanDepth_m",
                      "log_MaximumDepth_m",
                      "log_SurfaceArea_ha",
                      "log_Res_time",
                      "log_dynamic_ratio",
                      "Elevation_m",
                      "mean_chla")
responses <- c("chla_lag")
#responses <- c("Chla_ugL_EPI")#Nothing comes out cleanly
aic_calculator_onevar(for_mlr,responses,potential_drivers)
summary(lm(chla_lag~log_Res_time, data = for_mlr))

p_value = summary(lm(chla_lag~log_Res_time, data = for_mlr))$coefficients[,4][2]
if(p_value < 0.001){p_value_text = "< 0.001"}else{p_value_text = paste0("== ",as.character(round(p_value,3)))}
r2_value = summary(lm(chla_lag~log_Res_time, data = for_mlr))$r.squared
r2_value_text = round(r2_value,2)

vhod <- for_mlr%>%
  filter(!is.na(chla_lag))%>%
  ggplot(aes(x = Res_time,y=chla_lag))+
  scale_x_log10()+
  theme_bw()+
  geom_hline(yintercept = 0)+
  ylab("Coefficient for the effect\nof chl-a (t-1) on VHOD")+
  xlab("Residence time (d)")+
  geom_point(color = "grey50")+
  geom_smooth(method = "lm",color = "black")+
  annotate("text",x = 250, y = 0.73,
           label = paste0("R^2 == ",r2_value_text,"~~~p ", p_value_text), 
           parse = TRUE)

png("../Figures/Driver analysis AF.png", width = 7, height = 5, res = 1000, units = "in")
ggarrange(ggarrange(chla,
                    vhod,
                    labels = c("(a)","(b)"),
                    label.y = 1.02,
                    label.x = 0.02,
                    font.label = list(face = "bold")),
          ggarrange(demand_af,
                    ggplot()+geom_blank()+theme_minimal(),
                    labels = c("(c)", NA),
                    label.y = 1.02,
                    label.x = 0.02,
                    font.label = list(face = "bold")
                    ),
          nrow = 2
          )
dev.off()
```


Step 7: Before and after the first year of anoxia (figure)
```{r}
total <- with_temp%>%
  filter(!is.na(DO_mgL_HYPO))%>%
  group_by(LakeID)%>%
  filter(length(unique(Year))>=10)%>%
  ungroup()%>%
  summarize(length(unique(LakeID)))
total$`length(unique(LakeID))`

anoxic <- with_temp%>%
  filter(!is.na(DO_mgL_HYPO))%>%
  group_by(LakeID)%>%
  filter(length(unique(Year))>=10,
         max(DO_mgL_HYPO) < 1.8)%>%
  ungroup()
length(unique(anoxic$LakeID))
length(unique(anoxic$LakeID))/total$`length(unique(LakeID))`

oxic <- with_temp%>%
  filter(!is.na(DO_mgL_HYPO))%>%
  group_by(LakeID)%>%
  filter(length(unique(Year))>=10,
         min(DO_mgL_HYPO) > 1.8)%>%
  ungroup()
length(unique(oxic$LakeID))

switch <- with_temp%>%
  filter(!is.na(DO_mgL_HYPO))%>%
  group_by(LakeID)%>%
  filter(length(unique(Year))>=10,
         min(DO_mgL_HYPO) < 1.8,
         max(DO_mgL_HYPO) >= 1.8
         )%>%
  ungroup()
length(unique(switch$LakeID))

status_data <- with_temp%>%
  group_by(LakeID)%>%
  arrange(LakeID,Year)%>%
  filter(!is.na(DO_mgL_HYPO))%>%
  filter(sum(DO_mgL_HYPO < 1.8) >= 1,
         )%>%
  mutate(anoxic_date = min(Year[DO_mgL_HYPO < 1.8]),
         first_date = min(Year),
         end_date = max(Year),
         status = ifelse(Year>anoxic_date,"post-anoxia","pre-anoxia"),
         status = ifelse(Year==anoxic_date,"first anoxia",status),
         status = factor(status,levels = c("pre-anoxia","first anoxia","post-anoxia")))%>%
  filter(abs(Year-anoxic_date)<=5)%>%
  filter(sum(status=="pre-anoxia")>0,
         sum(status=="post-anoxia")>0,
         sum(status=="first anoxia")>0
         )%>%
  mutate(oxic = as.factor(DO_mgL_HYPO>1))

jpeg("../Figures/anoxia_status_do.jpg",res=300,width=3, height=3,units="in")
status_data%>%
  filter(!is.na(DO_mgL_HYPO),
         abs(Year-anoxic_date)<=1)%>%
  group_by(LakeID)%>%
  mutate(n = n())%>%
  filter(n == 3)%>%
  ggplot(aes(x=status,y=DO_mgL_HYPO))+
  geom_boxplot(outlier.shape = NA)+
  geom_jitter(width=.2, alpha = 0.2)+
  scale_color_viridis_c()+
  theme_bw()+
  #stat_compare_means()+
  theme(axis.title.x = element_blank(),
        legend.position = "none")+
  ylab("DO (mg/L)")
dev.off()

status_data_do <- status_data%>%
  filter(!is.na(DO_mgL_HYPO),
         abs(Year-anoxic_date) <= 1)%>%
  group_by(LakeID)%>%
  mutate(n = n())%>%
  filter(n == 3)

length(unique(status_data_do$LakeID))

friedman.test(y = status_data_do$DO_mgL_HYPO, groups = status_data_do$status, blocks = status_data_do$LakeID)
TukeyHSD(aov(lm(DO_mgL_HYPO~status+LakeID, data=status_data_do)))
```
